feat: extract iterative runtime helpers to separate module
Extract helper functions from iterative.py to iterative_runtime.py to reduce
complexity and enable deterministic unit testing. Moves _remaining_timeout,
_should_emit_progress, _append_overflow_hint, and _prune_with_debug to new
module while preserving all functionality and public API.
